requirements.txt
numpy==1.26.3
pandas==2.1.4
phe==1.5.0
gmpy2==2.1.5
crypto==1.4.1
cryptography==41.0.4
cmake==3.27.6
libsvm==3.23.0.4
lightgbm==4.2.0
pycryptodome==3.20.0
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
pyfate==2.0.0
fate-client==2.0.0
ucimlrepo==0.0.2